requirements.txt
Django==1.10.4
djangorestframework==3.5.3
Markdown==2.6.7
